+++ crates/nu-engine/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2018"
 
 [dependencies]
 nu-parser = { path = "../nu-parser" }
+nu-protocol = { path = "../nu-protocol" }
