+++ crates/nu-engine/src/lib.rs
@@ -1,7 +1,5 @@
 mod eval;
 mod state;
-mod value;
 
-pub use eval::{eval_block, eval_expression, eval_operator, ShellError};
+pub use eval::{eval_block, eval_expression, eval_operator};
 pub use state::{Stack, State};
-pub use value::Value;
