+++ crates/nu-parser/src/lib.rs
@@ -1,23 +1,14 @@
-mod declaration;
 mod errors;
 mod flatten;
 mod lex;
 mod lite_parse;
 mod parser;
 mod parser_state;
-mod signature;
-mod span;
 mod type_check;
 
-pub use declaration::Declaration;
 pub use errors::ParseError;
 pub use flatten::FlatShape;
 pub use lex::{lex, Token, TokenContents};
 pub use lite_parse::{lite_parse, LiteBlock};
-pub use parser::{
-    span, Block, Call, Expr, Expression, Import, Operator, Pipeline, Statement, SyntaxShape,
-    VarDecl,
-};
-pub use parser_state::{BlockId, DeclId, ParserDelta, ParserState, ParserWorkingSet, Type, VarId};
-pub use signature::{Flag, PositionalArg, Signature};
-pub use span::Span;
+pub use parser::{Block, Call, Expr, Expression, Import, Operator, Pipeline, Statement, VarDecl};
+pub use parser_state::{ParserDelta, ParserState, ParserWorkingSet};
