+++ crates/nu-parser/src/parser_state.rs
@@ -1,8 +1,8 @@
-use crate::{parser::Block, Declaration, Span};
+use crate::parser::Block;
 use core::panic;
-use std::{collections::HashMap, fmt::Display, slice::Iter};
+use nu_protocol::{BlockId, DeclId, Declaration, Span, Type, VarId};
+use std::{collections::HashMap, slice::Iter};
 
-#[derive(Debug)]
 pub struct ParserState {
     files: Vec<(String, usize, usize)>,
     file_contents: Vec<u8>,
@@ -11,49 +11,6 @@ pub struct ParserState {
     blocks: Vec<Block>,
     scope: Vec<ScopeFrame>,
 }
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum Type {
-    Int,
-    Float,
-    Bool,
-    String,
-    Block,
-    ColumnPath,
-    Duration,
-    FilePath,
-    Filesize,
-    List(Box<Type>),
-    Number,
-    Nothing,
-    Table,
-    Unknown,
-}
-
-impl Display for Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Type::Block => write!(f, "block"),
-            Type::Bool => write!(f, "bool"),
-            Type::ColumnPath => write!(f, "column path"),
-            Type::Duration => write!(f, "duration"),
-            Type::FilePath => write!(f, "filepath"),
-            Type::Filesize => write!(f, "filesize"),
-            Type::Float => write!(f, "float"),
-            Type::Int => write!(f, "int"),
-            Type::List(l) => write!(f, "list<{}>", l),
-            Type::Nothing => write!(f, "nothing"),
-            Type::Number => write!(f, "number"),
-            Type::String => write!(f, "string"),
-            Type::Table => write!(f, "table"),
-            Type::Unknown => write!(f, "unknown"),
-        }
-    }
-}
-
-pub type VarId = usize;
-pub type DeclId = usize;
-pub type BlockId = usize;
 
 #[derive(Debug)]
 struct ScopeFrame {
@@ -136,7 +93,7 @@ impl ParserState {
 
     pub fn print_decls(&self) {
         for decl in self.decls.iter().enumerate() {
-            println!("decl{}: {:?}", decl.0, decl.1);
+            println!("decl{}: {:?}", decl.0, decl.1.signature);
         }
     }
 
@@ -219,13 +176,11 @@ impl ParserState {
     }
 }
 
-#[derive(Debug)]
 pub struct ParserWorkingSet<'a> {
     permanent_state: &'a ParserState,
     pub delta: ParserDelta,
 }
 
-#[derive(Debug)]
 pub struct ParserDelta {
     files: Vec<(String, usize, usize)>,
     pub(crate) file_contents: Vec<u8>,
