+++ crates/nu-protocol/src/declaration.rs
@@ -1,6 +1,5 @@
 use crate::{BlockId, Signature};
 
-#[derive(Clone, Debug)]
 pub struct Declaration {
     pub signature: Box<Signature>,
     pub body: Option<BlockId>,
