+++ crates/nu-protocol/src/id.rs
@@ -1,0 +1,3 @@
+pub type VarId = usize;
+pub type DeclId = usize;
+pub type BlockId = usize;
